Enhance footer to make more sense
Before user can still go back to sign in page and join page when have login in the footer. Now, I added bound just like the nav bar when user log in the footer will change to users footer and when log out it go back to non user footer.